<refactoring> PostType 경로 변경
