+++ back/main_sever/src/posts/-comment/enum/post_type.ts
@@ -1,0 +1,4 @@
+export enum PostType {
+    QUEST = 'QUEST',
+    SB = 'SB',
+}
